requirements.txt
Django==5.1.*
djangorestframework==3.15.*
gunicorn==22.*
django-environ==0.11.*
docx2python==2.8.*
python-docx==1.1.*
diff-match-patch==20230430
whitenoise==6.7.*
weasyprint==62.*
pytest==8.*
pytest-django==4.*
